src/server/pfs/server/testing: split up and shard
