+++ src/server/pfs/server/testing/BUILD.bazel
@@ -1,22 +1,54 @@
-load("@rules_go//go:def.bzl", "go_test")
+load("@rules_go//go:def.bzl", "go_library", "go_test")
+
+# gazelle:go_test file
+go_test(
+    name = "fuzz_provenance_test",
+    size = "medium",
+    srcs = ["fuzz_provenance_test.go"],
+    data = glob(["testdata/**"]),
+    embed = [":testing"],
+    tags = ["cpu:8"],
+    deps = [
+        "//src/internal/client",
+        "//src/internal/dockertestenv",
+        "//src/internal/pctx",
+        "//src/internal/require",
+        "//src/internal/testpachd/realenv",
+        "//src/internal/testutil",
+        "//src/pfs",
+        "//src/server/pfs",
+        "@org_golang_google_protobuf//types/known/emptypb",
+    ],
+)
 
 go_test(
-    name = "testing_test",
-    size = "enormous",
-    srcs = [
-        "load_test.go",
-        "server_test.go",
-        "storage_test.go",
-    ],
+    name = "load_test",
+    srcs = ["load_test.go"],
     data = glob(["testdata/**"]),
-    pure = "on",
+    embed = [":testing"],
+    deps = [
+        "//src/internal/cmdutil",
+        "//src/internal/dockertestenv",
+        "//src/internal/pctx",
+        "//src/internal/require",
+        "//src/internal/testpachd/realenv",
+        "@org_golang_google_protobuf//types/known/emptypb",
+    ],
+)
+
+go_test(
+    name = "server_test",
+    srcs = ["server_test.go"],
+    data = glob(["testdata/**"]),
+    embed = [":testing"],
+    shard_count = 10,
+    tags = ["cpu:8"],
     deps = [
         "//src/auth",
         "//src/enterprise",
         "//src/internal/ancestry",
         "//src/internal/backoff",
         "//src/internal/client",
-        "//src/internal/cmdutil",
         "//src/internal/config",
         "//src/internal/dbutil",
         "//src/internal/dockertestenv",
@@ -45,8 +77,34 @@ go_test(
         "@org_golang_google_genproto_googleapis_rpc//errdetails",
         "@org_golang_google_grpc//codes",
         "@org_golang_google_grpc//status",
-        "@org_golang_google_protobuf//types/known/emptypb",
         "@org_golang_google_protobuf//types/known/wrapperspb",
         "@org_golang_x_sync//errgroup",
+    ],
+)
+
+go_test(
+    name = "storage_test",
+    size = "small",
+    srcs = ["storage_test.go"],
+    data = glob(["testdata/**"]),
+    embed = [":testing"],
+    deps = [
+        "//src/internal/dockertestenv",
+        "//src/internal/pctx",
+        "//src/internal/require",
+        "//src/internal/testpachd/realenv",
+        "//src/pfs",
+    ],
+)
+
+go_library(
+    name = "testing",
+    srcs = ["common.go"],
+    importpath = "github.com/pachyderm/pachyderm/v2/src/server/pfs/server/testing",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//src/internal/client",
+        "//src/pfs",
+        "//src/server/pfs",
     ],
 )
